RavenDB-22513 Fix modals render
